+++ walkers/src/extras/mod.rs
@@ -1,6 +1,9 @@
 //! Extra functionalities that can be used with the map.
+
+mod image;
 mod labeled_symbol;
 mod places;
 pub use crate::tiles::Texture;
+pub use image::Image;
 pub use labeled_symbol::{LabeledSymbol, LabeledSymbolStyle};
-pub use places::{Image, Places};
+pub use places::Places;
